refactor review billable action into a reusable method
